requirements/base.txt
awesome-slugify==1.6.5      # GPLv3
bpython==0.15				# MIT License
Django==1.9.10				# BSD License
django-appconf==1.0.2
django-braces==1.9.0		# BSD
django_compressor==2.0		# MIT
django-countries==3.4.1     # MIT
django-libsass==0.7		    # BSD
django-model-utils==2.5 	# BSD
djangorestframework==3.4.1  # BSD
# Dependency of djangorestframework
django-crispy-forms==1.6.0  # MIT
django-soapbox==1.3         # BSD
django-waffle==0.11.1		# BSD
pinax-announcements==2.0.4   # MIT
edx-auth-backends==0.5.0
edx-django-release-util==0.2.0
edx-i18n-tools==0.3.5
edx-rest-api-client>=1.5.0, <1.6.0  # Apache

libsass==0.11.1              # MIT

logutils==0.3.3				# BSD
python-social-auth==0.2.19

requests==2.10.0            # Apache 2.0

git+https://github.com/edx/django-lang-pref-middleware.git@0.1.0#egg=django-lang-pref-middleware
#git+https://github.com/edx/edx-analytics-data-api-client.git@0.10.0#egg=edx-analytics-data-api-client==0.10.0  # edX
git+https://github.com/CredoReference/edx-analytics-data-api-client.git@ec4edccfceeecda5e879a24a36cd24ce51716661#egg=edx-analytics-data-api-client==0.10.1  # edX
git+https://github.com/edx/opaque-keys.git@d45d0bd8d64c69531be69178b9505b5d38806ce0#egg=opaque-keys
# custom opaque-key implementations for ccx
git+https://github.com/jazkarta/ccx-keys.git@e6b03704b1bb97c1d2f31301ecb4e3a687c536ea#egg=ccx-keys

# Library to send info to Sentry
raven==5.32.0
